requirements.txt
# NewsBot 2.0 Intelligence System - Complete Requirements
# Advanced NLP Integration and Analysis Platform
# Updated for myenv3.10 compatibility with NewsBot 2.0 enhancements

# Core Data Processing
pandas>=2.0.3
numpy>=1.26.4
matplotlib>=3.7.2
seaborn>=0.13.1

# Machine Learning & Advanced Analytics
scikit-learn>=1.7.1
scipy>=1.13.1
xgboost>=2.1.0
lightgbm>=4.5.0

# Core NLP Libraries
nltk>=3.9.1
spacy>=3.8.4
textblob>=0.19.0
vaderSentiment>=3.3.2
flair>=0.14.0

# Advanced NLP & Language Models
transformers>=4.45.0
sentence-transformers>=3.1.0
torch>=2.4.0
datasets>=3.0.0
tokenizers>=0.20.0

# Topic Modeling & Unsupervised Learning
gensim>=4.3.3
pyldavis>=3.4.1

# Multilingual & Translation
langdetect>=1.0.9
deep-translator>=1.11.4
googletrans>=4.0.0rc1
google-cloud-translate>=3.17.0

# Semantic Search & Embeddings
faiss-cpu>=1.8.0
chromadb>=0.5.15
pinecone-client>=5.0.1

# Conversational AI & Intent Recognition
rasa>=3.6.20
dialogflow>=2.29.0
openai>=1.51.2

# Visualization & Interactive Dashboards
plotly>=6.2.0
wordcloud>=1.9.4
streamlit>=1.42.2
dash>=2.18.1
bokeh>=3.6.0

# Advanced Network Analysis
networkx>=3.4.2
pyvis>=0.3.2
graph-tool>=2.59

# Document Processing & Export
reportlab>=4.2.2
openpyxl>=3.1.5
python-docx>=1.1.2
pdfplumber>=0.11.4
pypdf2>=3.0.1

# Configuration & Environment Management
pyyaml>=6.0.2
python-dotenv>=1.0.1
configparser>=7.1.0

# Web Scraping & Data Acquisition
beautifulsoup4>=4.12.3
scrapy>=2.11.2
selenium>=4.25.0
aiohttp>=3.8.0
feedparser>=6.0.10

# Database & Storage
sqlalchemy>=2.0.35
redis>=5.1.1
pymongo>=4.10.1

# API & Web Framework
flask>=3.0.3
fastapi>=0.115.4
uvicorn>=0.32.0
celery>=5.4.0

# Testing & Quality Assurance
pytest>=8.3.3
pytest-cov>=5.0.0
black>=24.10.0
flake8>=7.1.1
mypy>=1.13.0

# Performance & Monitoring
memory-profiler>=0.61.0
psutil>=6.1.0
py-spy>=0.3.14

# Jupyter Environment
jupyter>=1.1.1
jupyterlab>=4.3.4
ipywidgets>=8.1.5
notebook>=7.3.2
nbconvert>=7.16.4

# Advanced Analysis (Enhanced)
umap-learn>=0.5.9
tqdm>=4.67.1
joblib>=1.4.2
dask>=2024.11.0

# Utility Libraries
requests>=2.32.3
pyarrow>=18.1.0
pydantic>=2.10.3
jinja2>=3.1.4
packaging>=24.2
typing-extensions>=4.12.2
python-dateutil>=2.9.0
regex>=2024.11.6
click>=8.1.7
rich>=13.9.4

# Security & Validation
cryptography>=43.0.3
certifi>=2024.8.30
urllib3>=2.2.3

# Optional: GPU Support (uncomment if using CUDA)
# torch-audio>=2.4.0
# torchaudio>=2.4.0
# torchvision>=0.19.0

# Optional: Additional Language Models (uncomment as needed)
# tensorflow>=2.17.0
# keras>=3.6.0
# jax>=0.4.35

# Development Tools
pre-commit>=4.0.1
isort>=5.13.2
bandit>=1.7.10 